+++ src/euroeval/dataset_configs/norwegian.py
@@ -297,36 +297,36 @@ GERLANGMOD_NN_CONFIG = DatasetConfig(
 )
 
 ZEBRA_PUZZLE_EASY_NB_CONFIG = DatasetConfig(
-    name="zebra-puzzle-easy-nb",
-    pretty_name="ZebraPuzzleEasy-nb",
-    source="EuroEval/zebra-puzzle-easy-nb",
+    name="zebra-puzzles-easy-nb",
+    pretty_name="ZebraPuzzlesEasy-nb",
+    source="EuroEval/zebra-puzzles-easy-nb",
     task=LOGIC,
     languages=[NORWEGIAN_BOKMÅL, NORWEGIAN],
     unofficial=True,
 )
 
 ZEBRA_PUZZLE_HARD_NB_CONFIG = DatasetConfig(
-    name="zebra-puzzle-hard-nb",
-    pretty_name="ZebraPuzzleHard-nb",
-    source="EuroEval/zebra-puzzle-hard-nb",
+    name="zebra-puzzles-hard-nb",
+    pretty_name="ZebraPuzzlesHard-nb",
+    source="EuroEval/zebra-puzzles-hard-nb",
     task=LOGIC,
     languages=[NORWEGIAN_BOKMÅL, NORWEGIAN],
     unofficial=True,
 )
 
 ZEBRA_PUZZLE_EASY_NN_CONFIG = DatasetConfig(
-    name="zebra-puzzle-easy-nn",
-    pretty_name="ZebraPuzzleEasy-nn",
-    source="EuroEval/zebra-puzzle-easy-nn",
+    name="zebra-puzzles-easy-nn",
+    pretty_name="ZebraPuzzlesEasy-nn",
+    source="EuroEval/zebra-puzzles-easy-nn",
     task=LOGIC,
     languages=[NORWEGIAN_NYNORSK, NORWEGIAN],
     unofficial=True,
 )
 
 ZEBRA_PUZZLE_HARD_NN_CONFIG = DatasetConfig(
-    name="zebra-puzzle-hard-nn",
-    pretty_name="ZebraPuzzleHard-nn",
-    source="EuroEval/zebra-puzzle-hard-nn",
+    name="zebra-puzzles-hard-nn",
+    pretty_name="ZebraPuzzlesHard-nn",
+    source="EuroEval/zebra-puzzles-hard-nn",
     task=LOGIC,
     languages=[NORWEGIAN_NYNORSK, NORWEGIAN],
     unofficial=True,